i18n_project/example_app/serializers.py
from rest_framework import serializers
from .models import Question

class QuestionSerializer(serializers.ModelSerializer):
    """
    Serializes the Question object
    """

    class Meta:
        model = Question
        fields = ['pk', 'question', 'answer']